requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
distlib==0.3.6
et-xmlfile==1.1.0
exceptiongroup==1.0.0
filelock==3.8.0
fonttools==4.37.4
h11==0.14.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pgeocode==0.3.0
Pillow==9.2.0
platformdirs==2.5.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-file==1.5.1
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tldextract==3.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.16.5
wsproto==1.2.0
